ipq807x: implement CPR based DVFS
Currently, DVFS for Acorn and Hawkeye is implemented by hardcoding the OPP
point voltages.
That works fine for Acorn and keeps it stable as simply the max voltages
for those points were used, however Hawkeye has an issue as it has much
higher clocks and more OPP points.
This can lead to unstable devices and weird crashes with memory errors
as the currently set voltages are not sufficent for all SoC-s as it does
not account for SoC binning at all.

So, lets for the time being pull in the Qualcomm CPR and APM drivers,
patch them to work in 5.15 and add the required DTS changes.

CPR and APM make sure that binning is took into account as well as other
information that is burned into eFuses during manufacturing as well
as making sure that memory controller, cache etc voltages also scale
with load.

Signed-off-by: Robert Marko <robimarko@gmail.com>
Signed-off-by: boos4721 <3.1415926535boos@gmail.com>